# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface TripInput {
  /** Driver's current address or city */
  current_location: string;
  /** Pickup address or city */
  pickup_location: string;
  /** Dropoff address or city */
  dropoff_location: string;
  /**
     * Hours already used in the current 70-hour/8-day cycle
     * @minimum 0
     * @maximum 70
     */
  current_cycle_used: number;
}

export interface Coordinates {
  lat: number;
  lng: number;
}

export type StopType = typeof StopType[keyof typeof StopType];


export const StopType = {
  current: 'current',
  pickup: 'pickup',
  dropoff: 'dropoff',
  fuel: 'fuel',
  rest: 'rest',
  break: 'break',
} as const;

export interface Stop {
  type: StopType;
  /** Human-readable place name */
  location: string;
  coordinates: Coordinates;
  /** ISO 8601 datetime */
  arrivalTime: string;
  /** Duration at stop in hours */
  duration: number;
  /**
     * Miles from trip origin
     * @nullable
     */
  distanceFromStart?: number | null;
}

export type LogSegmentStatus = typeof LogSegmentStatus[keyof typeof LogSegmentStatus];


export const LogSegmentStatus = {
  off_duty: 'off_duty',
  sleeper_berth: 'sleeper_berth',
  driving: 'driving',
  on_duty: 'on_duty',
} as const;

export interface LogSegment {
  /** Hour of day (0-24) when this segment starts */
  startHour: number;
  /** Hour of day (0-24) when this segment ends */
  endHour: number;
  status: LogSegmentStatus;
}

export interface DailyLog {
  /** ISO 8601 date string (YYYY-MM-DD) */
  date: string;
  /** 1-indexed day number within the trip */
  dayNumber: number;
  segments: LogSegment[];
  /** Total off-duty hours for the day */
  totalOffDuty: number;
  /** Total sleeper berth hours for the day */
  totalSleeperBerth: number;
  /** Total driving hours for the day */
  totalDriving: number;
  /** Total on-duty (not driving) hours for the day */
  totalOnDuty: number;
  /** Notable events for the day's log */
  remarks: string[];
}

export interface RouteGeometry {
  /** Array of [lng, lat] coordinate pairs forming the route polyline */
  coordinates: number[][];
}

export interface TripPlan {
  /** Total route distance in miles */
  totalDistanceMiles: number;
  /** Total driving time in hours */
  totalDrivingHours: number;
  /** Total elapsed trip time including stops */
  totalTripHours: number;
  /** ISO 8601 datetime of estimated arrival */
  estimatedArrival: string;
  /** ISO 8601 datetime of trip start */
  startTime: string;
  /** Remaining hours in the 70-hr/8-day cycle after trip */
  remainingCycleHours: number;
  fuelStopCount: number;
  restStopCount: number;
  stops: Stop[];
  dailyLogs: DailyLog[];
  routeGeometry: RouteGeometry;
}

export interface SavedTrip {
  id: number;
  /** ISO 8601 datetime */
  createdAt: string;
  currentLocation: string;
  pickupLocation: string;
  dropoffLocation: string;
  currentCycleUsed: number;
  plan: TripPlan;
}