# packages/api-zod/src/generated/types/coordinates.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Coordinates {
  lat: number;
  lng: number;
}