# packages/api-zod/src/generated/types/dailyLog.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogSegment } from './logSegment';

export interface DailyLog {
  /** ISO 8601 date string (YYYY-MM-DD) */
  date: string;
  /** 1-indexed day number within the trip */
  dayNumber: number;
  segments: LogSegment[];
  /** Total off-duty hours for the day */
  totalOffDuty: number;
  /** Total sleeper berth hours for the day */
  totalSleeperBerth: number;
  /** Total driving hours for the day */
  totalDriving: number;
  /** Total on-duty (not driving) hours for the day */
  totalOnDuty: number;
  /** Notable events for the day's log */
  remarks: string[];
}