# packages/api-zod/src/generated/types/logSegment.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LogSegmentStatus } from './logSegmentStatus';

export interface LogSegment {
  /** Hour of day (0-24) when this segment starts */
  startHour: number;
  /** Hour of day (0-24) when this segment ends */
  endHour: number;
  status: LogSegmentStatus;
}