# packages/api-zod/src/generated/types/routeGeometry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RouteGeometry {
  /** Array of [lng, lat] coordinate pairs forming the route polyline */
  coordinates: number[][];
}