# packages/api-zod/src/generated/types/savedTrip.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TripPlan } from './tripPlan';

export interface SavedTrip {
  id: number;
  /** ISO 8601 datetime */
  createdAt: string;
  currentLocation: string;
  pickupLocation: string;
  dropoffLocation: string;
  currentCycleUsed: number;
  plan: TripPlan;
}