# packages/api-zod/src/generated/types/stop.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Coordinates } from './coordinates';
import type { StopType } from './stopType';

export interface Stop {
  type: StopType;
  /** Human-readable place name */
  location: string;
  coordinates: Coordinates;
  /** ISO 8601 datetime */
  arrivalTime: string;
  /** Duration at stop in hours */
  duration: number;
  /**
     * Miles from trip origin
     * @nullable
     */
  distanceFromStart?: number | null;
}