# packages/api-zod/src/generated/types/tripInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TripInput {
  /** Driver's current address or city */
  current_location: string;
  /** Pickup address or city */
  pickup_location: string;
  /** Dropoff address or city */
  dropoff_location: string;
  /**
     * Hours already used in the current 70-hour/8-day cycle
     * @minimum 0
     * @maximum 70
     */
  current_cycle_used: number;
}