# packages/api-zod/src/generated/types/tripPlan.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * TruckRoute AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DailyLog } from './dailyLog';
import type { RouteGeometry } from './routeGeometry';
import type { Stop } from './stop';

export interface TripPlan {
  /** Total route distance in miles */
  totalDistanceMiles: number;
  /** Total driving time in hours */
  totalDrivingHours: number;
  /** Total elapsed trip time including stops */
  totalTripHours: number;
  /** ISO 8601 datetime of estimated arrival */
  estimatedArrival: string;
  /** ISO 8601 datetime of trip start */
  startTime: string;
  /** Remaining hours in the 70-hr/8-day cycle after trip */
  remainingCycleHours: number;
  fuelStopCount: number;
  restStopCount: number;
  stops: Stop[];
  dailyLogs: DailyLog[];
  routeGeometry: RouteGeometry;
}